Abstracting parse_dict pattern
